Aaaaand...I was right.
I floundered over this for a long time, thinking that I needed to slow
down React to see the letters progression, when in reality I just needed
to interpret the final result.

That wasn't clear, though, because my rectangle grid was too far spaced
apart to make the letters obvious.

Live and learn.