Progress and discovery
I generalized the "up_one" function to "move_one", which handles both up
and down moves, but then noticed I still have redundancy in my position
list because I am not sorting the generators, and when two or more
microchips are on the same floor, the corresponding generators can be
sorted in many different ways.

So for example, I have these two positions, which are the same:

m_code g_code
22244  13411
22244  31411

and actually, there are many more duplicates of those two, 24 in fact.