Added elevator moves down
Thought this would give me the solution, but when I gave my answer (26),
AoC said it was too low.  I will add a script to show each step of my
solution soon, so I can see whether I'm wrong (which is more likely) or
I've discovered a faster way to do it.