Tried to be tricky, but my answer was wrong
Unicode doesn't have 3 million+ consecutive characters, but it does have
consecutive characters from 1 to over 50000, so if I take two fairly
good-sized numbers that are relatively prime, such as 29999 and 30001,
and assign player numbers modulo those values, and compute the winner
for both, I should be able to determine the unique winning player.

I tried that, but it didn't work.